refactor(number-theory): isolate direct factorization contracts
